remove old tcl parser
