+++ sharding-core/src/main/resources/META-INF/parsing-rule-definition/postgresql/sql-statement-rule-definition.xml
@@ -10,11 +10,11 @@
     <sql-statement-rule context="alterIndex" sql-statement-class="io.shardingsphere.core.parsing.antlr.sql.statement.ddl.DDLStatement" extractor-rule-refs="indexNames" />
     <sql-statement-rule context="dropIndex" sql-statement-class="io.shardingsphere.core.parsing.antlr.sql.statement.ddl.DDLStatement" extractor-rule-refs="indexNames" />
     
-    <sql-statement-rule context="beginTransaction" sql-statement-class="io.shardingsphere.core.parsing.parser.sql.tcl.begin.BeginTransactionStatement" />
-    <sql-statement-rule context="commit" sql-statement-class="io.shardingsphere.core.parsing.parser.sql.tcl.commit.CommitStatement" />
-    <sql-statement-rule context="rollback" sql-statement-class="io.shardingsphere.core.parsing.parser.sql.tcl.rollback.RollbackStatement" />
-    <sql-statement-rule context="savepoint" sql-statement-class="io.shardingsphere.core.parsing.parser.sql.tcl.savepoint.SavepointStatement" />
-    <sql-statement-rule context="setTransaction" sql-statement-class="io.shardingsphere.core.parsing.parser.sql.tcl.set.transaction.SetTransactionStatement" />
+    <sql-statement-rule context="beginTransaction" sql-statement-class="io.shardingsphere.core.parsing.antlr.sql.statement.tcl.BeginTransactionStatement" />
+    <sql-statement-rule context="commit" sql-statement-class="io.shardingsphere.core.parsing.antlr.sql.statement.tcl.CommitStatement" />
+    <sql-statement-rule context="rollback" sql-statement-class="io.shardingsphere.core.parsing.antlr.sql.statement.tcl.RollbackStatement" />
+    <sql-statement-rule context="savepoint" sql-statement-class="io.shardingsphere.core.parsing.antlr.sql.statement.tcl.SavepointStatement" />
+    <sql-statement-rule context="setTransaction" sql-statement-class="io.shardingsphere.core.parsing.antlr.sql.statement.tcl.SetTransactionStatement" />
     
     <sql-statement-rule context="show" sql-statement-class="io.shardingsphere.core.parsing.parser.dialect.postgresql.statement.ShowStatement" extractor-rule-refs="showParam" />
     <sql-statement-rule context="setParam" sql-statement-class="io.shardingsphere.core.parsing.parser.dialect.postgresql.statement.SetParamStatement" />
