+++ sharding-core/src/test/resources/antlr_parser/mysql/tcl/set_transaction.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <parser-result-sets namespace="mysql.tcl.setTransaction">
-    <parser-result sql-case-id="set_transaction" tcl-actual-statement-class-type="io.shardingsphere.core.parsing.parser.sql.tcl.set.transaction.SetTransactionStatement" />
-    <parser-result sql-case-id="set_global_transaction" tcl-actual-statement-class-type="io.shardingsphere.core.parsing.parser.sql.tcl.set.transaction.SetTransactionStatement" />
-    <parser-result sql-case-id="set_transaction_read_only" tcl-actual-statement-class-type="io.shardingsphere.core.parsing.parser.sql.tcl.set.transaction.SetTransactionStatement" />
+    <parser-result sql-case-id="set_transaction" tcl-actual-statement-class-type="io.shardingsphere.core.parsing.antlr.sql.statement.tcl.SetTransactionStatement" />
+    <parser-result sql-case-id="set_global_transaction" tcl-actual-statement-class-type="io.shardingsphere.core.parsing.antlr.sql.statement.tcl.SetTransactionStatement" />
+    <parser-result sql-case-id="set_transaction_read_only" tcl-actual-statement-class-type="io.shardingsphere.core.parsing.antlr.sql.statement.tcl.SetTransactionStatement" />
 </parser-result-sets>
