+++ sharding-core/src/test/resources/antlr_parser/oracle/tcl/commit.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <parser-result-sets namespace="oracle.tcl.commit">
-    <parser-result sql-case-id="commit" tcl-actual-statement-class-type="io.shardingsphere.core.parsing.parser.sql.tcl.commit.CommitStatement" />
-    <parser-result sql-case-id="commit_with_comment" tcl-actual-statement-class-type="io.shardingsphere.core.parsing.parser.sql.tcl.commit.CommitStatement" />
-    <parser-result sql-case-id="commit_force" tcl-actual-statement-class-type="io.shardingsphere.core.parsing.parser.sql.tcl.commit.CommitStatement" />
+    <parser-result sql-case-id="commit" tcl-actual-statement-class-type="io.shardingsphere.core.parsing.antlr.sql.statement.tcl.CommitStatement" />
+    <parser-result sql-case-id="commit_with_comment" tcl-actual-statement-class-type="io.shardingsphere.core.parsing.antlr.sql.statement.tcl.CommitStatement" />
+    <parser-result sql-case-id="commit_force" tcl-actual-statement-class-type="io.shardingsphere.core.parsing.antlr.sql.statement.tcl.CommitStatement" />
 </parser-result-sets>
