+++ sharding-core/src/test/resources/antlr_parser/oracle/tcl/rollback.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <parser-result-sets namespace="oracle.tcl.rollback">
-    <parser-result sql-case-id="rollback" tcl-actual-statement-class-type="io.shardingsphere.core.parsing.parser.sql.tcl.rollback.RollbackStatement" />
-    <parser-result sql-case-id="rollback_to_savepoint" tcl-actual-statement-class-type="io.shardingsphere.core.parsing.parser.sql.tcl.rollback.RollbackStatement" />
-    <parser-result sql-case-id="rollback_force" tcl-actual-statement-class-type="io.shardingsphere.core.parsing.parser.sql.tcl.rollback.RollbackStatement" />
+    <parser-result sql-case-id="rollback" tcl-actual-statement-class-type="io.shardingsphere.core.parsing.antlr.sql.statement.tcl.RollbackStatement" />
+    <parser-result sql-case-id="rollback_to_savepoint" tcl-actual-statement-class-type="io.shardingsphere.core.parsing.antlr.sql.statement.tcl.RollbackStatement" />
+    <parser-result sql-case-id="rollback_force" tcl-actual-statement-class-type="io.shardingsphere.core.parsing.antlr.sql.statement.tcl.RollbackStatement" />
 </parser-result-sets>
