+++ sharding-core/src/test/resources/antlr_parser/oracle/tcl/savepoint.xml
@@ -1,4 +1,4 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <parser-result-sets namespace="oracle.tcl.savepoint">
-    <parser-result sql-case-id="savepoint" tcl-actual-statement-class-type="io.shardingsphere.core.parsing.parser.sql.tcl.savepoint.SavepointStatement" />
+    <parser-result sql-case-id="savepoint" tcl-actual-statement-class-type="io.shardingsphere.core.parsing.antlr.sql.statement.tcl.SavepointStatement" />
 </parser-result-sets>
