+++ sharding-core/src/test/resources/antlr_parser/postgre/tcl/commit.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <parser-result-sets namespace="postgre.tcl.commit">
-    <parser-result sql-case-id="commit" tcl-actual-statement-class-type="io.shardingsphere.core.parsing.parser.sql.tcl.commit.CommitStatement" />
-    <parser-result sql-case-id="commit_prepare" tcl-actual-statement-class-type="io.shardingsphere.core.parsing.parser.sql.tcl.commit.CommitStatement" />
+    <parser-result sql-case-id="commit" tcl-actual-statement-class-type="io.shardingsphere.core.parsing.antlr.sql.statement.tcl.CommitStatement" />
+    <parser-result sql-case-id="commit_prepare" tcl-actual-statement-class-type="io.shardingsphere.core.parsing.antlr.sql.statement.tcl.CommitStatement" />
 </parser-result-sets>
