+++ sharding-core/src/test/resources/antlr_parser/sqlserver/tcl/rollback.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <parser-result-sets namespace="sqlserver.tcl.rollback">
-    <parser-result sql-case-id="rollback" tcl-actual-statement-class-type="io.shardingsphere.core.parsing.parser.sql.tcl.rollback.RollbackStatement" />
-    <parser-result sql-case-id="rollback_with_name" tcl-actual-statement-class-type="io.shardingsphere.core.parsing.parser.sql.tcl.rollback.RollbackStatement" />
+    <parser-result sql-case-id="rollback" tcl-actual-statement-class-type="io.shardingsphere.core.parsing.antlr.sql.statement.tcl.RollbackStatement" />
+    <parser-result sql-case-id="rollback_with_name" tcl-actual-statement-class-type="io.shardingsphere.core.parsing.antlr.sql.statement.tcl.RollbackStatement" />
 </parser-result-sets>
