+++ sharding-core/src/test/resources/antlr_parser/sqlserver/tcl/set_transaction.xml
@@ -1,4 +1,4 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <parser-result-sets namespace="sqlserver.tcl.setTransaction">
-    <parser-result sql-case-id="set_transaction" tcl-actual-statement-class-type="io.shardingsphere.core.parsing.parser.sql.tcl.set.transaction.SetTransactionStatement" />
+    <parser-result sql-case-id="set_transaction" tcl-actual-statement-class-type="io.shardingsphere.core.parsing.antlr.sql.statement.tcl.SetTransactionStatement" />
 </parser-result-sets>
