+++ sharding-core/src/test/resources/parser/tcl.xml
@@ -1,12 +1,12 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <parser-result-sets>
-    <parser-result sql-case-id="setTransaction" tcl-actual-statement-class-type="io.shardingsphere.core.parsing.parser.sql.tcl.set.transaction.SetTransactionStatement" />
-    <parser-result sql-case-id="commit" tcl-actual-statement-class-type="io.shardingsphere.core.parsing.parser.sql.tcl.commit.CommitStatement" />
-    <parser-result sql-case-id="rollback" tcl-actual-statement-class-type="io.shardingsphere.core.parsing.parser.sql.tcl.rollback.RollbackStatement" />
-    <parser-result sql-case-id="savepoint" tcl-actual-statement-class-type="io.shardingsphere.core.parsing.parser.sql.tcl.savepoint.SavepointStatement" />
-    <parser-result sql-case-id="begin" tcl-actual-statement-class-type="io.shardingsphere.core.parsing.parser.sql.tcl.begin.BeginTransactionStatement" />
-    <parser-result sql-case-id="setAutoCommitOn" tcl-actual-statement-class-type="io.shardingsphere.core.parsing.parser.sql.tcl.set.autocommit.SetAutoCommitStatement" />
-    <parser-result sql-case-id="setAutoCommitOff" tcl-actual-statement-class-type="io.shardingsphere.core.parsing.parser.sql.tcl.set.autocommit.SetAutoCommitStatement" />
-    <parser-result sql-case-id="setImplicitTransactionsON" tcl-actual-statement-class-type="io.shardingsphere.core.parsing.parser.sql.tcl.set.autocommit.SetAutoCommitStatement" />
-    <parser-result sql-case-id="setImplicitTransactionsOff" tcl-actual-statement-class-type="io.shardingsphere.core.parsing.parser.sql.tcl.set.autocommit.SetAutoCommitStatement" />
+    <parser-result sql-case-id="setTransaction" tcl-actual-statement-class-type="io.shardingsphere.core.parsing.antlr.sql.statement.tcl.SetTransactionStatement" />
+    <parser-result sql-case-id="commit" tcl-actual-statement-class-type="io.shardingsphere.core.parsing.antlr.sql.statement.tcl.CommitStatement" />
+    <parser-result sql-case-id="rollback" tcl-actual-statement-class-type="io.shardingsphere.core.parsing.antlr.sql.statement.tcl.RollbackStatement" />
+    <parser-result sql-case-id="savepoint" tcl-actual-statement-class-type="io.shardingsphere.core.parsing.antlr.sql.statement.tcl.SavepointStatement" />
+    <parser-result sql-case-id="begin" tcl-actual-statement-class-type="io.shardingsphere.core.parsing.antlr.sql.statement.tcl.BeginTransactionStatement" />
+    <parser-result sql-case-id="setAutoCommitOn" tcl-actual-statement-class-type="io.shardingsphere.core.parsing.antlr.sql.statement.tcl.SetAutoCommitStatement" />
+    <parser-result sql-case-id="setAutoCommitOff" tcl-actual-statement-class-type="io.shardingsphere.core.parsing.antlr.sql.statement.tcl.SetAutoCommitStatement" />
+    <parser-result sql-case-id="setImplicitTransactionsON" tcl-actual-statement-class-type="io.shardingsphere.core.parsing.antlr.sql.statement.tcl.SetAutoCommitStatement" />
+    <parser-result sql-case-id="setImplicitTransactionsOff" tcl-actual-statement-class-type="io.shardingsphere.core.parsing.antlr.sql.statement.tcl.SetAutoCommitStatement" />
 </parser-result-sets>
